Add x86 support for the initrd= command line option
There is code in the core that allows specification of
the initrd as a physical address and size, but it is
not supported on all architectures.

For example, ARM today allows specification of physical
addresses and sizes for initrd in both device tree and
command line.  This is useful in cases where there is
no firmware file system structure in the FLASH.

One such situation occurs when we have reclaimed unused FLASH
space on UEFI systems by, e.g., taking it from the Managment
Engine. For example, on many systems, the ME is given half the
FLASH part; not only is 2.75M of an 8M part unused; but 10.75M
of a 16M part is unused. We can use this space to contain
an initrd, but need to tell Linux where it is.

This space is "raw": due to, e.g., UEFI limitations:
it can not be added to UEFI firmware volumes without rebuilding
UEFI from source or writing a UEFI device driver. We can reference it
only as a physical address and size.

At the same time, should we netboot a kernel or load it from
GRUB or syslinux, we want to have the option of not using
the physical address specification. Then, should we wish, it
is easy to boot the kernel and provide an initrd; or boot the
the kernel and let it use the initrd in FLASH. In practice this
has proven to be very helpful as we integrate Linux into FLASH
on x86.

Hence, the most flexible and convenient path is to enable the
initrd command line flag in a way that it is the last choice tried.

For example, on the DigitalLoggers Atomic Pi, we burn an image into
FLASH with a built-in command line which includes:
initrd=0xff968000,0x200000
which specifies a location and size.

We use this FLASH-based initrd only if it is not overridden by
other mechanisms.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>